# fix/feat: use joins instead of n+1/opening multiple connections