--- a/internal/storage/sql/db.go
+++ b/internal/storage/sql/db.go
@@ -31,11 +31,11 @@ func Open(cfg config.Config, opts ...Option) (*sql.DB, Driver, error) {
 
 	sql.SetMaxIdleConns(cfg.Database.MaxIdleConn)
 
-	if driver == SQLite {
-		sql.SetMaxOpenConns(1)
-	} else if cfg.Database.MaxOpenConn > 0 {
-		sql.SetMaxOpenConns(cfg.Database.MaxOpenConn)
-	}
+	// if driver == SQLite {
+	// 	sql.SetMaxOpenConns(1)
+	// } else if cfg.Database.MaxOpenConn > 0 {
+	sql.SetMaxOpenConns(cfg.Database.MaxOpenConn)
+	// }
 
 	if cfg.Database.ConnMaxLifetime > 0 {
 		sql.SetConnMaxLifetime(cfg.Database.ConnMaxLifetime)