… internal/storage/sql/common/segment.go