@GeorgeMac 2023-01-30
There is another approach to consider. We're doing the classic `N+1` but we could also do the Rails style `preload`.

Rails usually lets you pick between JOIN (`eager_load`) and the two queries style (`preload`).
Or `includes` which decides for you based on some heuristic I have forgotten.

The `preload` method just makes two queries.
The first is for the parent and the second uses `WHERE fk IN (id1, id2, ...)` to load the relation.