@markphelps 2023-01-31
Update: we chose to go with `option 1. Joins The Old Fashioned Way` and just continue doing the queries/results scanning by hand.

I continued the approach to the ListSegments and ListRules storage methods.

We are also now able to compare benchmarks added in #1307 (now on main) to those from this PR to see how using JOINS instead of multiple queries affects peformance/memory allocations.

Here is the results:

### Main Branch (no JOINS)

```
goos: linux
goarch: amd64
pkg: go.flipt.io/flipt/internal/storage/sql
cpu: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
BenchmarkListFlags/no-pagination-2         	      20	 289475937 ns/op	19884141 B/op	  677228 allocs/op
BenchmarkListFlags/pagination-limit-10-2   	    1681	   3557599 ns/op	  228084 B/op	    7661 allocs/op
BenchmarkListFlags/pagination-limit-25-2   	     763	   7836253 ns/op	  526125 B/op	   17818 allocs/op
BenchmarkListFlags/pagination-limit-100-2  	     204	  28974698 ns/op	 2016770 B/op	   68599 allocs/op
BenchmarkListFlags/pagination-limit-500-2  	      40	 145750363 ns/op	 9963442 B/op	  339415 allocs/op
BenchmarkListFlags/pagination-2            	      40	 147484753 ns/op	 9963565 B/op	  339424 allocs/op
BenchmarkListSegments/no-pagination-2      	       3	1891547016 ns/op	19284261 B/op	  697224 allocs/op
BenchmarkListSegments/pagination-limit-10-2         	     261	  23129548 ns/op	  221543 B/op	    7881 allocs/op
BenchmarkListSegments/pagination-limit-25-2         	     100	  53584245 ns/op	  510681 B/op	   18338 allocs/op
BenchmarkListSegments/pagination-limit-100-2        	      27	 199298428 ns/op	 1956430 B/op	   70620 allocs/op
BenchmarkListSegments/pagination-limit-500-2        	       6	 973445186 ns/op	 9662901 B/op	  349432 allocs/op
BenchmarkListSegments/pagination-2                  	       6	 974635037 ns/op	 9663716 B/op	  349444 allocs/op
```

## This Branch (Joins)

```
goos: linux
goarch: amd64
pkg: go.flipt.io/flipt/internal/storage/sql
cpu: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
BenchmarkListFlags/no-pagination-2         	      51	 115100076 ns/op	17045878 B/op	  695298 allocs/op
BenchmarkListFlags/pagination-limit-10-2   	    2778	   2183352 ns/op	   29532 B/op	    1031 allocs/op
BenchmarkListFlags/pagination-limit-25-2   	    2511	   2382021 ns/op	   54766 B/op	    2075 allocs/op
BenchmarkListFlags/pagination-limit-100-2  	    1878	   3163501 ns/op	  181777 B/op	    7290 allocs/op
BenchmarkListFlags/pagination-limit-500-2  	     802	   7504184 ns/op	  859892 B/op	   35098 allocs/op
BenchmarkListFlags/pagination-2            	     190	  31245636 ns/op	  860305 B/op	   35108 allocs/op
BenchmarkListSegments/no-pagination-2      	      44	 132385529 ns/op	16565384 B/op	  725297 allocs/op
BenchmarkListSegments/pagination-limit-10-2         	     319	  18577732 ns/op	   29003 B/op	    1064 allocs/op
BenchmarkListSegments/pagination-limit-25-2         	     322	  18671660 ns/op	   53517 B/op	    2153 allocs/op
BenchmarkListSegments/pagination-limit-100-2        	     306	  19551521 ns/op	  176936 B/op	    7593 allocs/op
BenchmarkListSegments/pagination-limit-500-2        	     246	  24178912 ns/op	  835893 B/op	   36601 allocs/op
BenchmarkListSegments/pagination-2                  	     134	  44454415 ns/op	  836265 B/op	   36611 allocs/op
```

If i'm reading it correctly, it looks like this approach results in anywhere from a 2-10x speed improvement AND uses anywhere from half to a tenth of the memory allocations!!

And also it fixes the SQLite contention issue, as we are now setting the `MaxOpenConns(1)` if running with SQLite.

Seems like a win/win..win.