@markphelps 2023-02-01
> This is awesome. There is one suggested quality of life change to config.

I moved the log warning to `internal/cmd/grpc` since we already have access to the logger there